Fix relative path to absolute path
