+++ src/App.tsx
@@ -1,8 +1,8 @@
+import "@/App.css";
+import Home from "@/pages/Home";
+import PortfolioMain from "@/pages/PortfolioMain";
+import { ModalStateProvider, ZDepthProvider } from "@/store/context";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import "./App.css";
-import Home from "./pages/Home";
-import PortfolioMain from "./pages/PortfolioMain";
-import { ModalStateProvider, ZDepthProvider } from "./store/context";
 
 function App() {
   return (
@@ -11,9 +11,7 @@ function App() {
         <ModalStateProvider>
           <Routes>
             <Route index element={<Home />} />
-            <Route path={"/page"} element={<PortfolioMain />}>
-              <Route path={":page"} element={<PortfolioMain />} />
-            </Route>
+            <Route path={"/page"} element={<PortfolioMain />} />
           </Routes>
         </ModalStateProvider>
       </ZDepthProvider>
