+++ src/lib/utils/image-loader-util.ts
@@ -7,7 +7,7 @@ export const getLocalPortfolioAssetPath = async (
   const images: string[] = [];
   for (let i = start; i <= (end || start); i++) {
     const { default: path } = await import(
-      `../assets/png/portfolio${i}.${extension || "png"}`
+      `@/assets/png/portfolio${i}.${extension || "png"}`
     );
     images.push(path);
   }
